+++ psycop/common/feature_generation/sequences/patient_loader.py
@@ -1,17 +1,13 @@
 import datetime
 from collections.abc import Sequence
+from dataclasses import dataclass
 from typing import Union
 
 import polars as pl
 
 from psycop.common.data_structures.patient import Patient
 from psycop.common.feature_generation.loaders.raw.load_demographic import birthdays
-from psycop.common.feature_generation.loaders.raw.load_ids import (
-    SplitName,
-    load_original_ids,
-)
 from psycop.common.feature_generation.sequences.event_loader import (
-    DiagnosisLoader,
     EventLoader,
 )
 from psycop.common.feature_generation.sequences.patient_slice_from_events import (
@@ -30,36 +26,28 @@ def keep_if_min_n_visits(
     return df
 
 
+@dataclass(frozen=True)
 class PatientLoader:
+    event_loaders: Sequence[EventLoader]
+    min_n_events: int | None = None
+    fraction: float = 1.0
+
     @staticmethod
     def load_date_of_birth_df() -> pl.DataFrame:
         df = pl.from_pandas(birthdays()).rename({"date_of_birth": "timestamp"})
 
         return df
 
-    @staticmethod
-    def get_split(
-        event_loaders: Sequence[EventLoader],
-        split: SplitName,
-        min_n_events: int | None = None,
-        fraction: float = 1.0,
-    ) -> Sequence[Patient]:
-        event_data = pl.concat([loader.load_events() for loader in event_loaders])
-        split_ids = (
-            pl.from_pandas(load_original_ids(split=split))
-            .sample(fraction=fraction)
-            .lazy()
-        )
+    def get_patients(self) -> Sequence[Patient]:
+        event_data = pl.concat([loader.load_events() for loader in self.event_loaders])
 
-        events_from_train = split_ids.join(event_data, on="dw_ek_borger", how="left")
-
-        if min_n_events:
-            events_from_train = keep_if_min_n_visits(
-                events_from_train,
-                n_visits=min_n_events,
+        if self.min_n_events:
+            event_data = keep_if_min_n_visits(
+                event_data,
+                n_visits=self.min_n_events,
             )
 
-        events_after_2013 = events_from_train.filter(
+        events_after_2013 = event_data.filter(
             pl.col("timestamp") > datetime.datetime(2013, 1, 1),
         )
 
@@ -69,11 +57,3 @@ class PatientLoader:
         )
 
         return unpacked_patients
-
-
-if __name__ == "__main__":
-    patients = PatientLoader.get_split(
-        event_loaders=[DiagnosisLoader()],
-        split=SplitName.TRAIN,
-        min_n_events=5,
-    )