+++ psycop/common/feature_generation/sequences/test_prediction_time_from_cohort.py
@@ -9,6 +9,10 @@ from psycop.common.feature_generation.sequences.prediction_times_from_cohort imp
 )
 from psycop.common.test_utils.str_to_df import str_to_pl_df
 
+from ...model_training_v2.trainer.preprocessing.steps.row_filter_split import (
+    FilterByEntityID,
+)
+
 
 class FakeCohortDefiner(CohortDefiner):
     @staticmethod
@@ -44,6 +48,7 @@ def test_polars_dataframe_to_dict():
             get_test_patient(patient_id=2),
             get_test_patient(patient_id=3),
         ],
+        split_filter=FilterByEntityID(split_ids=[1, 2, 3]),
     ).create_prediction_times(
         lookbehind=dt.timedelta(days=1),
         lookahead=dt.timedelta(days=1),