+++ psycop/common/feature_generation/text_models/encode_text_as_tfidf_scores.py
@@ -27,7 +27,7 @@ def encode_tfidf_values_to_df(
 
 if __name__ == "__main__":
     tfidf_model = load_text_model(
-        "tfidf_psycop_train_all_sfis_preprocessed_sfi_type_all_sfis_ngram_range_12_max_df_095_min_df_2_max_features_750.pkl",
+        "tfidf_psycop_train_all_sfis_preprocessed_sfi_type_all_sfis_ngram_range_12_max_df_09_min_df_2_max_features_750.pkl",
     )
 
     corpus = pl.from_pandas(
@@ -45,5 +45,5 @@ if __name__ == "__main__":
     TEXT_EMBEDDINGS_DIR.mkdir(exist_ok=True, parents=True)
     tfidf_notes.write_parquet(
         TEXT_EMBEDDINGS_DIR
-        / "text_tfidf_all_sfis_ngram_range_12_max_df_095_min_df_2_max_features_750.parquet",
+        / "text_tfidf_all_sfis_ngram_range_12_max_df_09_min_df_2_max_features_750.parquet",
     )