+++ psycop/common/model_training_v2/trainer/preprocessing/steps/test_cell_transformers.py
@@ -15,7 +15,7 @@ def test_bool_to_int():
         },
     )
 
-    transformed = BoolToInt().apply(pl.from_pandas(df))
+    transformed = BoolToInt().apply(pl.from_pandas(df).lazy()).collect()
     assert transformed.get_column("bool_col").to_list() == [1, 0, 1]
     assert transformed["int_col"].dtype == pl.Int64
     assert transformed["str_col"].dtype == pl.Utf8