+++ psycop/common/model_training_v2/trainer/preprocessing/steps/test_column_filters.py
@@ -14,7 +14,7 @@ def test_lookbehind_combination_filter():
         3,4,3,3
         4,3,4,1
         """,
-    )
+    ).lazy()
 
     from psycop.common.model_training_v2.loggers.terminal_logger import TerminalLogger
 
@@ -32,7 +32,7 @@ def test_regex_column_blacklist():
     df = str_to_pl_df(
         """pred_1, test_pred_1
         1, 2""",
-    )
+    ).lazy()
 
     filtered = RegexColumnBlacklist("pred_.*").apply(df)
     assert filtered.columns == [c for c in df.columns if not c.startswith("pred_")]
@@ -42,8 +42,8 @@ def test_temporal_column_filter():
     df = str_to_pl_df(
         """timestamp,pred_1,
                       2020-01-01,1,""",
-    )
-    filtered = TemporalColumnFilter().apply(df)
+    ).lazy()
+    filtered = TemporalColumnFilter().apply(df).lazy()
     assert filtered.columns == ["pred_1"]
 
 
@@ -51,7 +51,7 @@ def test_filter_within_subset():
     df = str_to_pl_df(
         """timestamp,pred_1,pred_2,
                       2020-01-01,1,1,""",
-    )
+    ).lazy()
 
     filtered = FilterColumnsWithinSubset(
         subset_rule="^pred_.+",