+++ psycop/projects/forced_admission_inpatient/model_eval/model_description/feature_importance/gain/gain_table.py
@@ -1,5 +1,7 @@
+import pandas as pd
 import polars as pl
 
+from psycop.common.global_utils.paths import OVARTACI_SHARED_DIR
 from psycop.common.model_training.data_loader.utils import (
     load_and_filter_split_from_cfg,
 )
@@ -14,10 +16,24 @@ from psycop.projects.forced_admission_inpatient.utils.pipeline_objects import (
 )
 
 
+def _load_vocabulary(
+    vocab_filename: str,
+) -> pd.DataFrame:
+    vocab_filepath = (
+        OVARTACI_SHARED_DIR / "text_models" / "vocabulary_lists" / vocab_filename
+    )
+
+    return pd.read_parquet(vocab_filepath)
+
+
 def generate_feature_importance_table(
     pipeline_run: ForcedAdmissionInpatientPipelineRun,
+    vocab_filename: str,
 ) -> pl.DataFrame:
     pipeline = pipeline_run.pipeline_outputs.pipe
+
+    # Import text model vocabulary
+    vocab = _load_vocabulary(vocab_filename)
 
     # Get feature importance scores
     feature_importances = pipeline.named_steps["model"].feature_importances_
@@ -29,13 +45,16 @@ def generate_feature_importance_table(
     )
     feature_names = [c for c in split_df.columns if "pred_" in c]
 
-    if "feature_selection" in pipeline["preprocessing"]:  # type: ignore
+    try:
+        pipeline[0]["feature_selection"]  # type: ignore
+
         feature_indices = pipeline["preprocessing"]["feature_selection"].get_support(  # type: ignore
             indices=True,
         )
         selected_feature_names = [feature_names[i] for i in feature_indices]
-    else:
-        selected_feature_names = feature_names
+
+    except KeyError:
+        selected_feature_names = feature_names  # type: ignore
 
     # Create a DataFrame to store the feature names and their corresponding gain
     feature_table = pl.DataFrame(
@@ -55,6 +74,14 @@ def generate_feature_importance_table(
     pd_df["index"] = pd_df["index"] + 1
     pd_df = pd_df.set_index("index")
 
+    # Map tfidf indices with actual ngrams from vocabulary
+    pd_df["Feature Name"][pd_df["Feature Name"].str.contains("tfidf")] = pd_df[
+        "Feature Name"
+    ][pd_df["Feature Name"].str.contains("tfidf")].str.replace(
+        r"\d+$",
+        lambda x: vocab.loc[int(x.group())]["Word"],
+    )
+
     with (
         pipeline_run.paper_outputs.paths.tables / "feature_importance_by_gain.html"
     ).open("w") as html_file:
@@ -67,4 +94,5 @@ def generate_feature_importance_table(
 if __name__ == "__main__":
     top_100_features = generate_feature_importance_table(
         pipeline_run=get_best_eval_pipeline(),
+        vocab_filename="vocab_tfidf_psycop_train_all_sfis_preprocessed_sfi_type_all_sfis_ngram_range_12_max_df_09_min_df_2_max_features_750.parquet",
     )